// src/app/core/api/model/templateResponseDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TemplateResponseDto { 
    framework: string;
    displayName: string;
    description: string;
    version: string;
    repo: string;
    repoUrl?: string;
    category: TemplateResponseDto.CategoryEnum;
    language: string;
    port: number;
    healthcheckPath: string;
    buildTool: string;
    isDefault: boolean;
    isDeprecated: boolean;
}
export namespace TemplateResponseDto {
    export const CategoryEnum = {
        Frontend: 'frontend',
        Backend: 'backend',
        Fullstack: 'fullstack',
        Static: 'static'
    } as const;
    export type CategoryEnum = typeof CategoryEnum[keyof typeof CategoryEnum];
}